chore(#572): Sonar 수정 — sort localeCompare + InMemoryKV 단일 fixture로 추출
