+++ backend/alarm-worker/src/__tests__/scheduled.test.ts
@@ -11,28 +11,7 @@ import {
 import { SeoulArrivalClient, type ArrivalEntry } from '../seoul';
 import { putTrip } from '../trips';
 import type { Env, Trip } from '../types';
-
-class InMemoryKV {
-  store = new Map<string, { value: string }>();
-  async get(key: string): Promise<string | null> {
-    return this.store.get(key)?.value ?? null;
-  }
-  async put(key: string, value: string): Promise<void> {
-    this.store.set(key, { value });
-  }
-  async delete(key: string): Promise<void> {
-    this.store.delete(key);
-  }
-  async list(options?: { prefix?: string }): Promise<{
-    keys: { name: string }[];
-    list_complete: boolean;
-    cursor: string;
-  }> {
-    const prefix = options?.prefix ?? '';
-    const keys = [...this.store.keys()].filter((k) => k.startsWith(prefix)).map((name) => ({ name }));
-    return { keys, list_complete: true, cursor: '' };
-  }
-}
+import { InMemoryKV } from './inMemoryKv';
 
 let apnsConfig: ApnsConfig;
